requirements-admin.txt
# Dreamina 管理后台依赖
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
requests>=2.28.0
pydantic>=2.0.0
